requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
deprecation==2.1.0
distro==1.9.0
docstring-parser==0.15
exceptiongroup==1.2.0
frozenlist==1.4.1
google-api-core==2.17.1
google-api-python-client==2.121.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
gotrue==2.4.1
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.25.2
idna==3.6
instructor==0.6.4
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
oauthlib==3.2.2
openai==1.13.3
packaging==23.2
postgrest==0.16.1
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.2
redis==5.0.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rq==1.16.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
storage3==0.7.0
StrEnum==0.4.15
supabase==2.4.0
supafunc==0.3.3
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typer==0.9.0
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
websockets==11.0.3
yarl==1.9.4
